+++ src/main/java/br/com/magnasistemas/cachacariaapi/service/ProdutoService.java
@@ -1,15 +1,16 @@
 package br.com.magnasistemas.cachacariaapi.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
-import javax.ws.rs.BeanParam;
+
+import org.modelmapper.ModelMapper;
 
 import br.com.magnasistemas.cachacariaapi.dao.ProdutoDAO;
 import br.com.magnasistemas.cachacariaapi.dto.ProdutoDTO;
 import br.com.magnasistemas.cachacariaapi.entity.Produto;
-import br.com.magnasistemas.cachacariaapi.settings.ProdutoMapper;
 
 @Stateless
 public class ProdutoService {
@@ -17,12 +18,14 @@ public class ProdutoService {
 	@Inject
 	private ProdutoDAO dao;
 
-	// public ProdutoMapper mapper;
+	ModelMapper modelmapper = new ModelMapper();
+	//ok com dto
+	public ProdutoDTO postProduto(Produto produto) {
 
-	/*
-	 * public ProdutoService(ProdutoMapper mapper) { super(); this.mapper = mapper;
-	 * }
-	 */
+		Produto prod = dao.postProduto(produto);
+
+		return modelmapper.map(prod, ProdutoDTO.class);
+	}
 
 	public List<Produto> findAllProduto() {
 
@@ -34,55 +37,62 @@ public class ProdutoService {
 		Produto produtoId = dao.findProdutoBYid(id);
 		return produtoId;
 	}
-	
-	public List<Produto> findWithName(String nome){
-		
-		return dao.findWithName(nome);
+
+	public List<ProdutoDTO> findWithName(String nome) {
+
+		List<Produto> produto = dao.findWithName(nome);
+		List<ProdutoDTO> produtosDTO = produto.stream().map(prods -> modelmapper.map(prods, ProdutoDTO.class))
+				.collect(Collectors.toList());
+		return produtosDTO ; 
 	}
 
-	public Produto postProduto(Produto produto) {
-		dao.postProduto(produto);
-		return produto;
-	}
-
-	/*
-	 * public ProdutoDTO postProduto(ProdutoDTO produtoDTO) {
-	 * dao.postProduto(mapper.paraEntity(produtoDTO)); return produtoDTO; }
-	 */
-
-	public Produto putProduto(long id, Produto produto) {
+	public ProdutoDTO putProduto(long id, Produto produto) {
 
 		Produto produtoAchado = dao.findProdutoBYid(id);
 		produtoAchado.setNome(produto.getNome());
 		produtoAchado.setPreco(produto.getPreco());
 		produtoAchado.setSabor(produto.getSabor());
 		dao.putProduto(produtoAchado);
-
-		return produtoAchado;
-	}
-
-	public Produto findProdutoBYid(long id) {
-
-		return dao.findProdutoBYid(id);
+		
+		ProdutoDTO produtoDTO;
+		produtoDTO = modelmapper.map(produtoAchado, ProdutoDTO.class);
+		return produtoDTO;
 	}
 
 	public void deleteProduto(Produto produto) {
-
-		dao.deleteProduto(produto);
+		
+		
+		if (produto != null) {
+			 dao.deleteProduto(produto);
+		}else {
+			 
+		}
+		
 	}
-	
-	
+
 }
 
 /*
- * public List<Produto> findAllProduto() { return dao.findAllProduto(); }
+ * public List<Produto> findAllProduto() {
+ * 
+ * return dao.findAllProduto(); }
  * 
  * public Produto findByIdProduto(long id) {
  * 
  * Produto produtoId = dao.findProdutoBYid(id); return produtoId; }
  * 
- * public Produto postProduto(Produto produto) { dao.postProduto(produto);
- * return produto; }
+ * public List<Produto> findWithName(String nome) {
+ * 
+ * return dao.findWithName(nome); }
+ * 
+ * 
+ * 
+ * 
+ * public ProdutoDTO postProduto(Produto produto) {
+ * 
+ * Produto prod = dao.postProduto(produto);
+ * 
+ * return modelmapper.map(prod, ProdutoDTO.class); }
  * 
  * public Produto putProduto(long id, Produto produto) {
  * 
@@ -93,11 +103,15 @@ public class ProdutoService {
  * 
  * return produtoAchado; }
  * 
- * public Produto findProdutoBYid(long id) {
- * 
- * return dao.findProdutoBYid(id); }
- * 
  * public void deleteProduto(Produto produto) {
  * 
  * dao.deleteProduto(produto); }
+ * 
+ * }
+ * 
+ * 
+ * public Produto postProduto(Produto produto) { dao.postProduto(produto);
+ * return produto; }
+ * 
+ * 
  */
